Edited services

+++ PaymentManagement.API/Models/Dtos/CreateInvoicePaymentDto.cs
@@ -9,7 +9,7 @@
         public int Cvv { get; set; }
         public int FlatId { get; set; }
         public int ExpenseId { get; set; }
-        public decimal InvoiceAmount { get; set; }
+        public double InvoiceAmount { get; set; }
     }
 }
