+++ PaymentManagement.API/Models/Entities/CreditCardInfo.cs
@@ -12,6 +12,6 @@ namespace PaymentManagement.API.Models.Entities
         public int ValidMonth { get; set; }
         public int ValidYear { get; set; }
         public int Cvv { get; set; }
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
     }
 }
